Use plugin instead of preference
